pytorch/feature: Record if parallel compile is enabled
This gets a bit messy, but this appears to be the best spot to make a
true / false decision.

Note that since we're looking at whether or not it's used, if the pool
doesn't warm up within the time it takes for a compile, we will mark the
feature use as false.

ghstack-source-id: 11420b32ff68283e6275128d5976f1ad9aab8543
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141074