[inductor] Don't specialize `split` on `sizes` parameter.
Fix: #139936

This PR modifies the lowering of `split` operation, so that it won't generate guards,
specializing on the sizes parameter. Instead, it specializes on the number of output
tensors being generated (i.e. function of the size of the base tensor, and the sizes
parameter).

As a result, operations such as `chunk` (whose number of output tensors usually is
constant given a static chunk number) won't trigger recompiles when varying the size of
the base tensor.

ghstack-source-id: 2ae0f8c4a960ffb8b94583a7b6ee89dd9989a7c7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141077